Final implementation of ticking UI

+++ app/src/main/java/com/bars/timer_test/data/TimerRepository.kt
@@ -7,11 +7,11 @@ import javax.inject.Inject
 
 class TimerRepository @Inject constructor(private val timerLocalDataStore: ITimerDataStore) : ITimerRepository {
 
-    override suspend fun loadTimer(context: Context): UInt {
+    override suspend fun loadTimer(context: Context): Long {
         return timerLocalDataStore.loadTimer(context)
     }
 
-    override fun saveTimer(timer: UInt, context: Context) {
+    override fun saveTimer(timer: Long, context: Context) {
         timerLocalDataStore.saveTimer(timer, context)
     }
 }
