+++ app/src/main/java/com/bars/timer_test/data/source/ITimerDataStore.kt
@@ -3,6 +3,6 @@ package com.bars.timer_test.data.source
 import android.content.Context
 
 interface ITimerDataStore {
-    suspend fun loadTimer(context: Context): UInt
-    fun saveTimer(timer: UInt, context: Context)
+    suspend fun loadTimer(context: Context): Long
+    fun saveTimer(timer: Long, context: Context)
 }
