+++ app/src/main/java/com/bars/timer_test/domain/repository/ITimerRepository.kt
@@ -3,6 +3,6 @@ package com.bars.timer_test.domain.repository
 import android.content.Context
 
 interface ITimerRepository {
-    suspend fun loadTimer(context: Context): UInt
-    fun saveTimer(timer: UInt, context: Context)
+    suspend fun loadTimer(context: Context): Long
+    fun saveTimer(timer: Long, context: Context)
 }
